UTXO转账完成

+++ part3/BLC/UTXO.go
@@ -2,6 +2,6 @@ package BLC
 
 type UTXO struct {
 	TxHash []byte
-	Index  int
+	Index  int64
 	Output *TXOutput
 }
